12R.4: Fix bookmark identity to use stable IDs
- Changed bookmarks from HashSet<usize> to HashSet<String>
- toggle_bookmark now takes Tab instead of tab_index
- is_bookmarked now takes Tab instead of tab_index
- get_bookmarked_tabs renamed to get_bookmarked_tab_ids returning Vec<String>
- Updated session capture/restore to work with stable IDs
- Removed app.current_tab as usize from bookmark paths

+++ crates/slapper/src/tui/app/runner.rs
@@ -257,7 +257,7 @@ where
                         }
                     }
                     (KeyModifiers::CONTROL, KeyCode::Char('b')) if app.mode == InputMode::Normal => {
-                        app.toggle_bookmark(app.current_tab as usize);
+                        app.toggle_bookmark(app.current_tab);
                     }
                     _ if app
                         .get_command_palette()
